Initial commit.
All assest were regenerated and should match the (deprecated)
live version that used to be deployed to branch www of repo transltr.

I humbly believe it is easier to manage two distinct repositories
stored in the same parent directory. I personally found that switching
from R to HTML in the same repo was annoying.